Added command to create payment, refactoring

+++ Api/BccPay.Core.Infrastructure/PaymentModels/NetsNodes/Item.cs
@@ -38,10 +38,10 @@
 
         [Required]
         [JsonPropertyName("netTotalAmount")]
-        public int NetTotalAmount { get; set; }
+        public decimal NetTotalAmount { get; set; }
 
         [Required]
         [JsonPropertyName("grossTotalAmount")]
-        public int GrossTotalAmount { get; set; }
+        public decimal GrossTotalAmount { get; set; }
     }
 }
